Proxy class is now universal data exchange class
Provide a convert(DST *, const SRC *, size_t) function for the two types
you want to build a ConversionProxy from and values will be
automagically converted if needed. If GLVertexAttribute has a proxy set
now with setProxy() it will not duplicate the attribute data, but pull
it in and/or convert it on the fly when it is bound.